changed the parentScreen according to the reviews, refactored rectangleButton
