+++ .gitignore
@@ -12,3 +12,4 @@ highScores.txt
 .pacman.ui.screens
 .project
 .vscode
+.project
