+++ src/test/java/com/thoughtworks/pacman/ui/screens/WinScreenTest.java
@@ -25,7 +25,7 @@ public class WinScreenTest {
         Game game = new Game();
         WinScreen winScreen = new WinScreen(game);
 
-        winScreen.setReturnToMainMenu(true);
+        winScreen.setStartGame(true);
 
         assertThat(winScreen.getNextScreen(), instanceOf(IntroScreen.class));
     }
